Project Title: Clustering and Segmentation in the US
In this project, I used data from the 2019 Survey of Consumer Finances. First, I identified households that have a hard time getting credit. Then yI built a model to segment these households into subgroups. Finally, I created an interactive web app to share my work.

This project is an example of unsupervised learning, specifically clustering. It can be used in commercial contexts for marketing or customer segmentation or in sociological contexts to study social stratification. Python and scikit-learn were used in this project.